+++ firefly_reports/__init__.py
@@ -1,4 +1,4 @@
 """Firefly reports."""
 
-__version__ = "0.3.1"
+__version__ = "0.4.0"
 __description__ = "Generate reports using Firefly's API."